requirements.txt
flask
werkzeug
pillow
scikit-learn
